requirements.txt
python-dotenv
fastapi[standard]
pydantic
SQLAlchemy
psycopg2-binary
openai